[chore] remove utils internal package
Moving funcs in use to different package
